refactor: fix WSL build issues, drop using-directives, convert cv_utils to class
Fix all compilation issues found in a fresh WSL2/GCC 11.4 build, including the -Werror quality gate: bytetrack_benchmark detector.tomlt() -> init(), kalman_filter tomltiate() -> initiate(), link glog into common (base64_unittest undefined google::LogMessage), mark vendored 3rd_party include dirs as SYSTEM in web_console, suppress vendored lapjv.cpp unused-parameter warnings, and fix sign-compare / maybe-uninitialized / missing-field-initializers / unused-variable issues across detectors, diffusion samplers, lightglue, SAM, server base impl and TRT/ONNX helpers.

Replace every using namespace directive in src/ and test/ with explicit using-declarations for the names actually used, and move cv_utils.h helpers into class CvUtils with static members so call sites use using jinq::common::CvUtils; + CvUtils::XXX(...) (a using-declaration cannot name a namespace).

Also include web-console TOML config parsing, CI repo-cleanliness check, StatusCode stream operator, include hygiene, Testing/ gitignore and related test adjustments.

Testing: full Release build and all 18 unit-test targets compile with 0 errors; unit tests 16/17 pass. cv_utils_unittest keeps two pre-existing functional assertion failures (calc_iou zero-area expectation, colorize_sam_mask_max_label).

+++ src/apps/server/classification/mobilenetv2_classification_server.cpp
@@ -52,7 +52,7 @@ int main(int argc, char** argv) {
     auto server = create_mobilenetv2_cls_server("mobilenetv2_cls_server");
         auto status = server->init(config);
         if (status != jinq::common::StatusCode::OK) {
-            LOG(ERROR) << "server init failed, status: " << std::to_string(status);
+            LOG(ERROR) << "server init failed, status: " << std::to_string(static_cast<int>(status));
             return -1;
         }
         if (server->start(host.c_str(), static_cast<unsigned short>(port)) == 0) {
